+++ theme/js/components/dataset/search-result.vue
@@ -1,5 +1,15 @@
 <template>
-  <article class="fr-pt-5v fr-pb-6v fr-px-1w border-bottom border-default-grey fr-enlarge-link" :style="style">
+  <article class="fr-my-3w fr-p-3w border border-default-grey fr-enlarge-link" :style="style">
+    <div class="absolute top-0 fr-grid-row fr-grid-row--middle fr-mt-n3v" v-if="private || archived">
+      <p class="fr-badge fr-badge--mention-grey fr-mr-1w" v-if="private">
+        <span class="fr-icon-lock-line" aria-hidden="true"></span>
+        {{ $t('Private') }}
+      </p>
+      <p class="fr-badge fr-badge--mention-grey" v-if="archived">
+        <span class="fr-icon-archive-line" aria-hidden="true"></span>
+        {{ $t('Archived') }}
+      </p>
+    </div>
     <div class="fr-grid-row fr-grid-row--gutters fr-grid-row--top">
       <div class="fr-col-auto">
         <div class="logo">
@@ -27,15 +37,9 @@
             {{ title }}
             <small v-if="acronym">{{ acronym }}</small>
           </a>
-          <span
-            v-if="private"
-            class="badge grey-300 fr-ml-1w"
-          >
-            {{ $t('Private') }}
-          </span>
         </h4>
         <span class="not-enlarged" v-if="organization || owner">
-          {{ $t('From') }} 
+          {{ $t('From') }}
           <a :href="organization.page" v-if="organization">
               <OrganizationNameWithCertificate :organization="organization" />
           </a>
@@ -137,7 +141,6 @@
 
 <script>
 import { defineComponent, computed, ComputedRef } from "vue";
-import lock from "bundle-text:svg/private.svg";
 import useLicense from "../../composables/useLicense";
 import useOwnerName from "../../composables/useOwnerName";
 import Avatar from "../discussions/avatar.vue";
@@ -159,6 +162,10 @@ export default defineComponent({
   inheritAttrs: false,
   props: {
     acronym: String,
+    archived: {
+      type: Boolean,
+      default: false,
+    },
     description: {
       type: String,
       required: true,
@@ -178,7 +185,10 @@ export default defineComponent({
       type: String,
       required: true,
     },
-    private: Boolean,
+    private: {
+      type: Boolean,
+      default: false,
+    },
     quality: {
       type: Object,
       required: true,
@@ -209,7 +219,6 @@ export default defineComponent({
     const license = useLicense(props.license);
     return {
       license,
-      lock,
       ownerName,
     };
   }