+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Fix Api Representation for media after CaptchEtat integration [#173](https://github.com/etalab/udata-front/pull/173)
+- Add archived and private badges to dataset search results and update card style [#170](https://github.com/etalab/udata-front/pull/170)
 
 ## 3.0.0 (2022-11-14)